drop webpack in favor of esbuild
We must still use TypeScript to build the code
that runs under node, because esbuild somehow
eats the puppeteer import (something about the
behavior of esModuleInterop?).

But for the web-facing bit, we can go direct from
TS to a bundle quickly enough that we can drop a
bunch of build machinery.  Delightful.

Also the package.json is a workaround for
https://github.com/evanw/esbuild/issues/274